CBOR: properly convert linearization data before doing a read/write

+++ Projects/GME_Binary/main/CBOR_CAREL.c
@@ -2505,8 +2505,8 @@ C_RES parse_write_values(c_cborreqrdwrvalues cbor_wv)
 				 hr_to_read.info.dim = cbor_wv.dim;
 				 hr_to_read.info.bitposition = cbor_wv.pos;
 				 hr_to_read.info.len = cbor_wv.len;
-				 hr_to_read.info.linA = atoi((C_SCHAR*)cbor_wv.a);
-				 hr_to_read.info.linB = atoi((C_SCHAR*)cbor_wv.b);
+				 hr_to_read.info.linA = atof((C_SCHAR*)cbor_wv.a);
+				 hr_to_read.info.linB = atof((C_SCHAR*)cbor_wv.b);
 				 hr_to_read.info.flag.byte = cbor_wv.flags.byte;
 
 				 // read the actual data
@@ -2579,8 +2579,8 @@ C_RES parse_read_values(c_cborreqrdwrvalues* cbor_rv){
 		hr_to_read.info.dim = cbor_rv->dim;
 		hr_to_read.info.bitposition = cbor_rv->pos;
 		hr_to_read.info.len = cbor_rv->len;
-		hr_to_read.info.linA = atoi((C_SCHAR*)cbor_rv->a);
-		hr_to_read.info.linB = atoi((C_SCHAR*)cbor_rv->b);
+		hr_to_read.info.linA = atof((C_SCHAR*)cbor_rv->a);
+		hr_to_read.info.linB = atof((C_SCHAR*)cbor_rv->b);
 		hr_to_read.info.flag.byte = cbor_rv->flags.byte;
 
 		result = PollEngine__Read_HR_IR_Req(cbor_rv->func, hr_to_read.info.Addr, hr_to_read.info.dim ,(void*)&hr_to_read.c_value.value);
